pyproject.toml
[tool.poetry]
name = "imdb-chatbot-svc"
version = "0.1.0"
description = ""
authors = ["raakhal-rapolu <pandu3011@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.40.2"
requests = "^2.32.3"
flask = "^3.1.0"
taskipy = "^1.14.1"
flask-cors = "^5.0.0"
flask-restx = "^1.3.0"
flask-swagger-ui = "^4.11.1"
chromadb = "^0.5.23"
llama-index = "^0.12.5"
python-dotenv = "^1.0.1"
llama-index-llms-groq = "^0.3.1"
groq = "^0.16.0"
sentence-transformers = "^3.4.1"
google-generativeai = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.taskipy.tasks]
imdb-svc-backend = { cmd = "python app.py" }
imdb-svc-frontend = { cmd = "streamlit run streamlit_ui2.py" }

